# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Callytics API",
    "description": "음성 분석 및 텍스트 처리 마이크로서비스 API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "개발 서버"
    },
    {
      "url": "http://localhost:8001",
      "description": "오디오 처리 서버"
    },
    {
      "url": "http://localhost:8002",
      "description": "텍스트 분석 서버"
    },
    {
      "url": "http://localhost:8003",
      "description": "데이터베이스 서버"
    }
  ],
  "paths": {
    "/preprocess": {
      "post": {
        "tags": [
          "Audio Processing"
        ],
        "summary": "오디오 전처리",
        "description": "오디오 파일을 전처리하여 분석에 적합한 형태로 변환",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AudioInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/SuccessResponse"
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/enhance": {
      "post": {
        "tags": [
          "Audio Processing"
        ],
        "summary": "오디오 품질 향상",
        "description": "노이즈 제거 및 음성 품질 향상",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AudioInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/SuccessResponse"
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/segment": {
      "post": {
        "tags": [
          "Audio Processing"
        ],
        "summary": "화자 분리",
        "description": "오디오에서 화자를 분리하고 세그먼트 생성",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AudioInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/SuccessResponse"
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AudioInput": {
        "type": "object",
        "properties": {
          "audio_data": {
            "type": "string",
            "description": "Base64 인코딩된 오디오 데이터"
          },
          "sample_rate": {
            "type": "integer",
            "description": "샘플링 레이트 (Hz)",
            "default": 16000
          },
          "format": {
            "type": "string",
            "description": "오디오 포맷",
            "enum": [
              "wav",
              "mp3",
              "flac"
            ],
            "default": "wav"
          }
        },
        "required": [
          "audio_data"
        ]
      },
...
    },
    "responses": {
      "SuccessResponse": {
        "description": "성공 응답",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "message": {
                  "type": "string",
                  "example": "처리 완료"
                },
                "data": {
                  "type": "object"
                }
              },
              "required": [
                "status"
              ]
            }
          }
        }
      },
      "ErrorResponse": {
        "description": "오류 응답",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "message": {
                  "type": "string",
                  "example": "처리 실패"
                },
                "error_code": {
                  "type": "string"
                },
                "details": {
                  "type": "object"
                }
              },
              "required": [
                "status",
                "message"
              ]
            }
          }
        }
      }
    }
  }
}